completed AGTSP
