--- AG/333.py
@@ -1,2 +1,0 @@
-a = 1 << 21
-print(a)
